test: fold the optimization characterization tests into the suite
The TDD scaffolding built for the xarray/dask optimization work was never
meant to ship as part of the library. The tests that pin real behavior move
into the modules covering the code they exercise, renamed to read as
ordinary regression tests, and the laziness guards move with them so that
eager compute cannot be quietly reintroduced.

`is_lazy` and `assert_lazy` are the only shared helper the survivors need,
and they now live in `tests/conftest.py`.

The baseline-relative performance contracts are dropped. Dask graph sizes,
task-execution counts and call counts encode one machine's shape of the
graph rather than the behavior of the code, and the wall-clock harness they
backstopped is no longer tracked in the repo.
