--- benchmarks/baseline.json
@@ -1,13 +1,0 @@
-{
-  "atmospheric_river_mask": 0.1071187480010849,
-  "atmospheric_river_mask_forecast": 1.19473,
-  "atmospheric_river_pipeline": 1.1396167470011278,
-  "cape": 0.02068979199975729,
-  "center_of_mass": 0.06922011300048325,
-  "convert_init_time_to_valid_time": 4.018879179999203,
-  "daily_min": 0.7787926500022877,
-  "landfall_detection": 0.15211,
-  "reduce_dataarray": 0.07069252399742254,
-  "region_mask": 0.012333730002865195,
-  "tc_spatial_mask": 0.60733
-}
