+++ tests/test_utils.py
@@ -10,6 +10,7 @@ import sparse
 import xarray as xr
 
 from extremeweatherbench import utils
+from tests.conftest import assert_lazy
 
 
 @pytest.mark.parametrize(
@@ -2044,3 +2045,468 @@ class TestIsValidLandfall:
         # This should return True because init_time IS in coords
         # (even though it's not a dimension)
         assert utils.is_valid_landfall(da) is True
+
+
+def _forecast_dataset(n_init, n_lead, chunk=True, lead_unit=None):
+    """Forecast-shaped dataset with init_time and lead_time dimensions."""
+    lead = np.arange(0, n_lead * 6, 6)
+    if lead_unit is not None:
+        lead = pd.to_timedelta(lead, unit=lead_unit)
+    ds = xr.Dataset(
+        {
+            "t": (
+                ["init_time", "lead_time", "latitude"],
+                np.arange(float(n_init * n_lead * 4)).reshape(n_init, n_lead, 4),
+            )
+        },
+        coords={
+            "init_time": pd.date_range("2020-01-01", periods=n_init, freq="D"),
+            "lead_time": lead,
+            "latitude": np.linspace(0, 3, 4),
+        },
+    )
+    return ds.chunk({"init_time": 1}) if chunk else ds
+
+
+def _spatial_dataarray(n_time=6, n_lat=8, n_lon=8, chunk=True):
+    """Small (valid_time, latitude, longitude) array for reduction tests."""
+    rng = np.random.default_rng(3)
+    da = xr.DataArray(
+        rng.uniform(280.0, 310.0, (n_time, n_lat, n_lon)),
+        dims=["valid_time", "latitude", "longitude"],
+        coords={
+            "valid_time": pd.date_range("2023-01-01", periods=n_time, freq="6h"),
+            "latitude": np.linspace(30.0, 30.0 + n_lat - 1, n_lat),
+            "longitude": np.linspace(-120.0, -120.0 + n_lon - 1, n_lon),
+        },
+    )
+    return da.chunk({"valid_time": 1}) if chunk else da
+
+
+def _daily_series(n_days, timesteps_per_day=4, chunk=True, drop_last=0):
+    """Hourly-ish series over whole days, optionally with a truncated last day."""
+    n_steps = n_days * timesteps_per_day - drop_last
+    freq = f"{24 // timesteps_per_day}h"
+    valid_time = pd.date_range("2021-06-01", periods=n_steps, freq=freq)
+    values = np.arange(float(n_steps))
+    da = xr.DataArray(values, dims=["valid_time"], coords={"valid_time": valid_time})
+    if chunk:
+        da = da.chunk({"valid_time": timesteps_per_day})
+    return da
+
+
+def _daily_min_via_map(da, time_resolution_hours):
+    """The groupby().map() formulation this phase replaces."""
+    return da.groupby("valid_time.dayofyear").map(
+        utils.min_if_all_timesteps_present,
+        time_resolution_hours=time_resolution_hours,
+    )
+
+
+def _landfall_dataarray(n_init=64, chunk=True, all_nan=False):
+    """Landfall-shaped DataArray indexed by init_time."""
+    values = np.full(n_init, np.nan) if all_nan else np.arange(float(n_init))
+    da = xr.DataArray(
+        values,
+        dims=["init_time"],
+        coords={"init_time": pd.date_range("2021-08-20", periods=n_init, freq="6h")},
+    )
+    return da.chunk({"init_time": 8}) if chunk else da
+
+
+def _sparse_dataarray():
+    """DataArray backed by a sparse.COO array."""
+    import sparse
+
+    data = sparse.COO(
+        coords=np.array([[0, 1, 2], [1, 2, 0]]),
+        data=np.array([1.0, 2.0, 3.0]),
+        shape=(4, 4),
+    )
+    return xr.DataArray(
+        data,
+        dims=["latitude", "longitude"],
+        coords={
+            "latitude": np.linspace(0.0, 3.0, 4),
+            "longitude": np.linspace(10.0, 13.0, 4),
+        },
+    )
+
+
+class TestTimeConversionOutput:
+    """Pins the exact values the lead-time reshape produces.
+
+    Expected values are hand-derived from the init/lead grid rather than
+    computed by the function under test.
+    """
+
+    def test_overlapping_valid_times_land_on_a_shared_axis(self):
+        """Two inits a day apart with 0 h and 24 h leads share a valid time.
+
+        init 2020-01-01 at lead 24h and init 2020-01-02 at lead 0h both land
+        on 2020-01-02, so the union axis has three entries, not four.
+        """
+        ds = xr.Dataset(
+            {"t": (["init_time", "lead_time"], [[0.0, 1.0], [2.0, 3.0]])},
+            coords={
+                "init_time": pd.to_datetime(["2020-01-01", "2020-01-02"]),
+                "lead_time": pd.to_timedelta([0, 24], unit="h"),
+            },
+        )
+
+        result = utils.convert_init_time_to_valid_time(ds)
+
+        assert list(result.valid_time.values) == list(
+            pd.to_datetime(["2020-01-01", "2020-01-02", "2020-01-03"])
+        )
+        actual = result["t"].transpose("lead_time", "valid_time").values
+        expected = np.array(
+            [
+                [0.0, 2.0, np.nan],  # lead 0h: inits 01-01, 01-02, then nothing
+                [np.nan, 1.0, 3.0],  # lead 24h: nothing, then inits 01-01, 01-02
+            ]
+        )
+        np.testing.assert_array_equal(actual, expected)
+
+    def test_integer_lead_times_are_added_as_nanoseconds(self):
+        """Integer lead_time is added directly to datetime64, i.e. as ns.
+
+        This pins existing behavior. Callers that mean hours must pass a
+        timedelta; the unit must not be silently reinterpreted.
+        """
+        ds = xr.Dataset(
+            {"t": (["init_time", "lead_time"], [[0.0, 1.0]])},
+            coords={
+                "init_time": pd.to_datetime(["2020-01-01"]),
+                "lead_time": [0, 24],
+            },
+        )
+
+        result = utils.convert_init_time_to_valid_time(ds)
+
+        assert list(result.valid_time.values) == [
+            np.datetime64("2020-01-01T00:00:00.000000000"),
+            np.datetime64("2020-01-01T00:00:00.000000024"),
+        ]
+
+    def test_init_time_is_retained_as_a_two_dimensional_coordinate(self):
+        """init_time survives as a (lead_time, valid_time) coord, NaT if absent."""
+        ds = xr.Dataset(
+            {"t": (["init_time", "lead_time"], [[0.0, 1.0], [2.0, 3.0]])},
+            coords={
+                "init_time": pd.to_datetime(["2020-01-01", "2020-01-02"]),
+                "lead_time": pd.to_timedelta([0, 24], unit="h"),
+            },
+        )
+
+        result = utils.convert_init_time_to_valid_time(ds)
+
+        init_coord = result.init_time.transpose("lead_time", "valid_time").values
+        assert init_coord.shape == (2, 3)
+        assert init_coord[0, 0] == np.datetime64("2020-01-01")
+        assert init_coord[0, 1] == np.datetime64("2020-01-02")
+        assert pd.isna(init_coord[0, 2])
+        assert pd.isna(init_coord[1, 0])
+        assert init_coord[1, 1] == np.datetime64("2020-01-01")
+        assert init_coord[1, 2] == np.datetime64("2020-01-02")
+
+    def test_non_dimension_coordinates_are_carried_through_and_masked(self):
+        """A bool (init_time, lead_time) coord survives the reshape.
+
+        ForecastBase.subset_data_to_case attaches valid_time_mask this way,
+        so the reshape has to carry it across without dropping it.
+        """
+        ds = xr.Dataset(
+            {"t": (["init_time", "lead_time"], [[0.0, 1.0], [2.0, 3.0]])},
+            coords={
+                "init_time": pd.to_datetime(["2020-01-01", "2020-01-02"]),
+                "lead_time": pd.to_timedelta([0, 24], unit="h"),
+                "valid_time_mask": (
+                    ["init_time", "lead_time"],
+                    np.array([[False, True], [True, True]]),
+                ),
+            },
+        )
+
+        result = utils.convert_init_time_to_valid_time(ds)
+
+        assert "valid_time_mask" in result.coords
+        mask = result.valid_time_mask.transpose("lead_time", "valid_time").values
+        assert not bool(mask[0, 0])
+        assert bool(mask[0, 1])
+        assert bool(mask[1, 1])
+
+    def test_chunked_input_gives_the_same_values_as_unchunked(self):
+        chunked = utils.convert_init_time_to_valid_time(_forecast_dataset(4, 3))
+        unchunked = utils.convert_init_time_to_valid_time(
+            _forecast_dataset(4, 3, chunk=False)
+        )
+        xr.testing.assert_allclose(chunked.compute(), unchunked)
+
+    def test_round_trip_back_to_init_time_recovers_the_original_values(self):
+        ds = _forecast_dataset(4, 3, chunk=False)
+        forward = utils.convert_init_time_to_valid_time(ds)
+        back = utils.convert_valid_time_to_init_time(forward["t"])
+
+        recovered = back.transpose("init_time", "lead_time", "latitude")
+        np.testing.assert_allclose(
+            recovered.sel(init_time=ds.init_time).values, ds["t"].values
+        )
+
+
+class TestReduceDataArrayLaziness:
+    """A spatial mean must not be a compute barrier by default.
+
+    Computing inside every reduction breaks the graph into one eager step per
+    call site, so a metric that reduces a forecast and a target pays two
+    separate passes over the data instead of one fused pass.
+    """
+
+    def test_string_method_reduction_stays_lazy(self):
+        da = _spatial_dataarray()
+        result = utils.reduce_dataarray(
+            da, method="mean", reduce_dims=["latitude", "longitude"], skipna=True
+        )
+        assert_lazy(result)
+
+    def test_callable_reduction_stays_lazy(self):
+        da = _spatial_dataarray()
+        result = utils.reduce_dataarray(
+            da, method=np.nanmean, reduce_dims=["latitude", "longitude"]
+        )
+        assert_lazy(result)
+
+    def test_compute_true_is_still_available_for_data_dependent_indexing(self):
+        da = _spatial_dataarray()
+        result = utils.reduce_dataarray(
+            da,
+            method="mean",
+            reduce_dims=["latitude", "longitude"],
+            compute=True,
+            skipna=True,
+        )
+        assert isinstance(result.data, np.ndarray)
+
+
+class TestReduceDataArrayOutput:
+    """Pins the reduced values, which laziness must not change."""
+
+    def test_lazy_and_eager_reductions_agree(self):
+        da = _spatial_dataarray()
+        lazy = utils.reduce_dataarray(
+            da, method="mean", reduce_dims=["latitude", "longitude"], skipna=True
+        )
+        eager = utils.reduce_dataarray(
+            da,
+            method="mean",
+            reduce_dims=["latitude", "longitude"],
+            compute=True,
+            skipna=True,
+        )
+        xr.testing.assert_allclose(lazy.compute(), eager)
+
+    def test_mean_matches_a_hand_computed_column(self):
+        da = _spatial_dataarray(n_time=2, n_lat=2, n_lon=2, chunk=False)
+        da.values = np.array([[[1.0, 2.0], [3.0, 4.0]], [[10.0, 20.0], [30.0, 40.0]]])
+        result = utils.reduce_dataarray(
+            da.chunk({"valid_time": 1}),
+            method="mean",
+            reduce_dims=["latitude", "longitude"],
+            skipna=True,
+        )
+        np.testing.assert_allclose(result.compute().values, [2.5, 25.0])
+
+
+class TestDailyMinOverCompleteDays:
+    """Pins the per-day minimum, taken over complete days only."""
+
+    def test_result_is_still_lazy(self):
+        da = _daily_series(n_days=8)
+        result = utils.daily_min_over_complete_days(da, 6.0)
+        assert_lazy(result, "the daily minimum should not materialize on construction")
+
+    def test_a_dask_backed_series_does_not_hit_the_flox_grouped_path(self):
+        # flox 0.10.8 raises "Cannot call len() on object with unknown chunk
+        # size" for any dask-backed grouped reduction against dask 2025.12, so
+        # the daily minimum must reach its answer without one.
+        da = _daily_series(n_days=8)
+        result = utils.daily_min_over_complete_days(da, 6.0).compute()
+        np.testing.assert_allclose(result.values, np.arange(8) * 4.0)
+
+    def test_complete_days_match_the_groupby_map_result(self):
+        da = _daily_series(n_days=5, chunk=False)
+        got = utils.daily_min_over_complete_days(da, 6.0)
+        expected = _daily_min_via_map(da, 6.0)
+        np.testing.assert_allclose(got.values, expected.values)
+
+    def test_an_incomplete_day_is_nan(self):
+        da = _daily_series(n_days=5, chunk=False, drop_last=1)
+        got = utils.daily_min_over_complete_days(da, 6.0)
+
+        assert np.isnan(got.values[-1]), "the truncated last day should be NaN"
+        np.testing.assert_allclose(got.values[:-1], [0.0, 4.0, 8.0, 12.0])
+
+    def test_lead_time_is_preserved_for_forecast_shaped_input(self):
+        base = _daily_series(n_days=4, chunk=False)
+        da = base.expand_dims(lead_time=[0, 6, 12]).copy()
+        got = utils.daily_min_over_complete_days(da, 6.0)
+
+        assert got.dims == ("lead_time", "dayofyear")
+        np.testing.assert_allclose(got.sel(lead_time=6).values, [0.0, 4.0, 8.0, 12.0])
+
+    def test_days_stay_matched_to_their_labels_across_a_year_boundary(self):
+        # dayofyear restarts below the days already seen, so the day labels no
+        # longer arrive in time order.
+        da = _daily_series(n_days=400, chunk=False)
+        got = utils.daily_min_over_complete_days(da, 6.0)
+        expected = _daily_min_via_map(da, 6.0)
+
+        np.testing.assert_array_equal(got.dayofyear.values, expected.dayofyear.values)
+        np.testing.assert_allclose(got.values, expected.values)
+
+    def test_all_days_incomplete_gives_nan_on_the_day_axis(self):
+        da = _daily_series(n_days=1, chunk=False, drop_last=1)
+        got = utils.daily_min_over_complete_days(da, 6.0)
+        assert got.dims == ("dayofyear",)
+        assert np.isnan(got.values).all()
+
+
+class TestDensifyDoesNotMutateTheCaller:
+    """Densifying returns a new array rather than rewriting the caller's."""
+
+    def test_the_original_dataarray_stays_sparse(self):
+        import sparse
+
+        da = _sparse_dataarray()
+        utils.maybe_densify_dataarray(da)
+
+        assert isinstance(da.data, sparse.COO), (
+            "the caller's array was densified in place; maybe_densify_dataarray "
+            "should return a new array instead"
+        )
+
+    def test_a_dataset_holding_the_array_stays_sparse(self):
+        import sparse
+
+        ds = xr.Dataset({"t": _sparse_dataarray()})
+        utils.maybe_densify_dataarray(ds["t"])
+
+        assert isinstance(ds["t"].data, sparse.COO), (
+            "densifying a variable pulled out of a dataset rewrote the "
+            "dataset's own copy"
+        )
+
+    def test_the_returned_array_is_dense_with_the_same_values(self):
+        da = _sparse_dataarray()
+        densified = utils.maybe_densify_dataarray(da)
+
+        assert not hasattr(densified.data, "todense")
+        np.testing.assert_allclose(densified.values, da.data.todense())
+
+    def test_coordinates_and_dims_survive(self):
+        da = _sparse_dataarray()
+        densified = utils.maybe_densify_dataarray(da)
+
+        assert densified.dims == da.dims
+        np.testing.assert_array_equal(densified.latitude.values, da.latitude.values)
+        np.testing.assert_array_equal(densified.longitude.values, da.longitude.values)
+
+    def test_a_dense_array_is_passed_straight_through(self):
+        da = xr.DataArray(np.arange(6.0).reshape(2, 3), dims=["y", "x"])
+        assert utils.maybe_densify_dataarray(da) is da
+
+
+class TestIsValidLandfallEdgeCases:
+    """Pins the verdict is_valid_landfall gives for edge-case inputs."""
+
+    @pytest.mark.parametrize("chunk", [False, True])
+    @pytest.mark.parametrize(
+        "kwargs, expected",
+        [
+            ({}, True),
+            ({"all_nan": True}, False),
+        ],
+    )
+    def test_verdict_is_unchanged(self, chunk, kwargs, expected):
+        da = _landfall_dataarray(n_init=16, chunk=chunk, **kwargs)
+        assert utils.is_valid_landfall(da) is expected
+
+    def test_none_and_scalar_are_still_invalid(self):
+        assert utils.is_valid_landfall(None) is False
+        assert utils.is_valid_landfall(xr.DataArray(np.nan)) is False
+
+    def test_missing_init_time_is_invalid(self):
+        da = xr.DataArray([1.0, 2.0], dims=["landfall"])
+        assert utils.is_valid_landfall(da) is False
+
+    def test_a_single_real_value_among_nans_is_valid(self):
+        da = _landfall_dataarray(n_init=16, chunk=True, all_nan=True)
+        da = da.copy()
+        da[5] = 3.0
+        assert utils.is_valid_landfall(da) is True
+
+
+class TestCoordinateOrderHandling:
+    """Coordinate order is read off the index, not via DataArrays."""
+
+    @pytest.mark.parametrize("descending", [False, True])
+    def test_region_mask_handles_both_latitude_directions(self, descending):
+        from extremeweatherbench import regions
+
+        latitudes = np.linspace(-90, 90, 73)
+        if descending:
+            latitudes = latitudes[::-1]
+        ds = xr.Dataset(
+            {"t": (["latitude", "longitude"], np.zeros((73, 144)))},
+            coords={"latitude": latitudes, "longitude": np.linspace(-180, 177.5, 144)},
+        )
+        region = regions.BoundingBoxRegion(
+            latitude_min=20.0,
+            latitude_max=50.0,
+            longitude_min=-130.0,
+            longitude_max=-100.0,
+        )
+        masked = region.mask(ds)
+
+        kept = masked.latitude.values
+        assert kept.size > 0, "the latitude slice came back empty"
+        assert kept.min() >= 20.0 and kept.max() <= 50.0
+
+    @pytest.mark.parametrize("hours", [1, 3, 6, 12])
+    def test_temporal_resolution_is_unchanged(self, hours):
+        ds = xr.Dataset(
+            {"t": ("valid_time", np.zeros(10))},
+            coords={
+                "valid_time": pd.date_range("2021-01-01", periods=10, freq=f"{hours}h")
+            },
+        )
+        assert utils.determine_temporal_resolution(ds) == float(hours)
+
+    def test_temporal_resolution_of_a_single_timestep_is_none(self):
+        ds = xr.Dataset(
+            {"t": ("valid_time", np.zeros(1))},
+            coords={"valid_time": pd.date_range("2021-01-01", periods=1)},
+        )
+        assert utils.determine_temporal_resolution(ds) is None
+
+    def test_temporal_resolution_takes_the_finest_of_mixed_spacings(self):
+        times = pd.to_datetime(
+            ["2021-01-01T00", "2021-01-01T06", "2021-01-01T18", "2021-01-02T00"]
+        )
+        ds = xr.Dataset(
+            {"t": ("valid_time", np.zeros(4))}, coords={"valid_time": times}
+        )
+        assert utils.determine_temporal_resolution(ds) == 6.0
+
+    def test_temporal_resolution_works_without_a_valid_time_index(self):
+        ds = xr.Dataset(
+            {"t": ("step", np.zeros(5))},
+            coords={
+                "step": np.arange(5),
+                "valid_time": (
+                    "step",
+                    pd.date_range("2021-01-01", periods=5, freq="3h"),
+                ),
+            },
+        )
+        assert utils.determine_temporal_resolution(ds) == 3.0
